[packaging] Remove bluez dependency and add bluez CONFIG option. Fixes JB#36287
QtSystems is LGPL/GPL dual-licensed. Having a dependency on the
GPL-licensed BlueZ means this distribution of QtSystems is
automatically also GPL-licensed and cannot be LGPL, which is not
desirable.

To prevent accidental building with any installed bluez libs,
also added config option so that QtSystems will not pull them in unless
CONFIG+=bluez is set.

Conflicts:
	rpm/qtsystems.spec